+++ python/paddle/fluid/tests/unittests/dygraph_to_static/CMakeLists.txt
@@ -46,7 +46,6 @@ endif()
 
 # Todo: support windows and apple
 if(WIN32 OR APPLE)
-  list(REMOVE_ITEM TEST_OPS test_resnet_prim)
   list(REMOVE_ITEM TEST_OPS test_cinn_prim_forward)
 endif()
 
@@ -72,7 +71,6 @@ set_tests_properties(test_transformer PROPERTIES TIMEOUT 200)
 set_tests_properties(test_bmn PROPERTIES TIMEOUT 120)
 #set_tests_properties(test_mnist PROPERTIES TIMEOUT 120)
 set_tests_properties(test_build_strategy PROPERTIES TIMEOUT 120)
-set_tests_properties(test_resnet_prim PROPERTIES TIMEOUT 120)
 
 if(NOT WIN32)
   set_tests_properties(test_resnet_v2 PROPERTIES TIMEOUT 120)
